net/ice/base: fix integer types in comparisons
- This patch resolves violations raised by Windows CodeQL
  tool related to comparison-with-wider-type error.

- Added data type casting to resolve these errors.

Fixes: f3202a097f12 ("net/ice/base: add ACL module")
Fixes: c9e37832c95f ("net/ice/base: rework on bit ops")
Fixes: 18a6dd522289 ("net/ice/base: improve switch advanced rule")
Fixes: 88824213be8a ("net/ice/base: enable RSS for PFCP/L2TP/ESP/AH")
Fixes: d647871c50ab ("net/ice/base: fix bitmap set function")
Fixes: 1082f786547e ("net/ice/base: support DCB")
Fixes: 93e84b1bfc92 ("net/ice/base: add basic Tx scheduler")
Fixes: fed0c5ca5f19 ("net/ice/base: support programming a new switch recipe")
Cc: stable@dpdk.org

Acked-by: Bruce Richardson <bruce.richardson@intel.com>

+++ drivers/net/intel/ice/base/ice_switch.c
@@ -7437,7 +7437,7 @@ ice_fill_fv_word_index(struct ice_hw *hw, struct LIST_HEAD_TYPE *fv_list,
 			struct ice_fv_word *pr;
 			bool found = false;
 			u16 mask;
-			u8 j;
+			u16 j;
 
 			pr = &rg->r_group.pairs[i];
 			mask = rg->r_group.mask[i];
